Update README.md & install.sh, add rulefiles

+++ install.sh
@@ -3,16 +3,16 @@
 #########################################
 # Got r00t o are u Amy Acker?
 #########################################
-if [[ $(/usr/bin/id -u) -ne 0 ]]; then
-    echo "rule-update needs root privileges or be Amy Acker.  Please try again using sudo."
-    exit
+if [ "$EUID" -ne 0 ]
+    then echo "rule-update needs root privileges or be Amy Acker.  Please try again using sudo."
+    exit 1
 fi
 
 #########################################
 # Instalar dependencias
 #########################################
 sudo apt-get install incron
-echo "No olvides añadir el daemon incrond al arranque del sistema. https://github.com/andoniaf/snort-Rulez#como-añado-el-daemon-incrond " >> LEEME.txt
+echo "No olvides añadir el daemon incrond al arranque del sistema. https://github.com/andoniaf/snort-Rulez#como-a%C3%B1ado-el-daemon-incrond " >> LEEME.txt
 
 #########################################
 # Copiar archivos de la aplicación a la ruta correspondiente
@@ -23,25 +23,47 @@ echo "Esta ruta es la que se usará para acceder a la aplicación desde el naveg
 echo "Se creara una carpeta con el nombre de SnortRulez."
 read -p "Ruta : " PATH_PHP
 #Creamos la carpeta y copiamos en ella todos los archivos
-mkdir $PATH_PHP/SnortRulez
-cp ./* $PATH_PHP/SnortRulez
+if ! [ -d $PATH_PHP ]; then
+        while true; do
+                echo "La ruta no existe. Introduce una ruta válida."
+                read -p "Ruta : " PATH_PHP
+                if [ -d $PATH_PHP ]; then
+                break
+                fi
+        done
+fi
+mkdir -p $PATH_PHP/SnortRulez
+cp -r ./* $PATH_PHP/SnortRulez
+chown www-data:www-data $PATH_PHP/test.rules
+chown www-data:www-data $PATH_PHP/easy.rules
+chown www-data:www-data $PATH_PHP/custom.rules
 
 #Crear los enlaces de los archivos rules a su ruta correspondiente
 echo "Escriba la ruta donde estan guardadas las reglas."
 echo "Si estas usando SecurityOnion la ruta probablemente sera: /etc/nsm/rules"
 read -p "Ruta : " PATH_RULES
+if ! [ -d $PATH_RULES ]; then
+        while true; do
+                echo "La ruta no existe. Introduce una ruta válida."
+                read -p "Ruta : " PATH_RULES
+                if [ -d $PATH_RULES ]; then
+                break
+                fi
+        done
+fi
+
 ln $PATH_PHP/SnortRulez/test.rules $PATH_RULES/test.rules
 ln $PATH_PHP/SnortRulez/easy.rules $PATH_RULES/easy.rules
 ln $PATH_PHP/SnortRulez/custom.rules $PATH_RULES/custom.rules
 
 #########################################
-# Crear e importar la BD                ¡¡¡¡¡¡¡¡¡Una vez verificado que funciona cambiar el nombre de la BD
+# Crear e importar la BD
 #########################################
-BDName="SnortPrueba"
-STTY_SAVE=$(stty -g)
-stty -echo
+BDName="SnortRulez"
 echo "----------------------------------------------------------------------"
 read -p "Escriba el nombre de usuario de la BBDD [p.e: root]: " BDUser
+STTY_SAVE=$(stty -g)
+stty -echo
 echo -n "Introduce la contraseña del usuario de la BBDD: "
 read DBADMIN_SECRET_PASSWD
 stty $STTY_SAVE
